# openapi.yaml
openapi: 3.0.3
info:
  title: Beacon Backend API
  version: 1.1.0
paths:
...
  /sos:
    post:
      summary: Create SOS thread and notify friends
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSosRequest"
      responses:
        "200":
          description: SOS created
        "400":
          $ref: "#/components/responses/BadRequest"
  /sos/{sosId}/status:
    patch:
      summary: Set terminal SOS outcome
      security:
        - bearerAuth: []
      parameters:
        - name: sosId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SosStatusUpdateRequest"
      responses:
        "200":
          description: SOS status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SosStatusUpdateResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    CreateSosRequest:
      type: object
      required: [category]
      properties:
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        address:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        category:
          type: string
          enum: [medical, fire, violence, unknown]
    SosStatusUpdateRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [cancelled, safe]
        resolved_at:
          type: string
          format: date-time
          nullable: true
        source:
          type: string
          enum: [android]
          nullable: true
    SosStatusUpdateResponse:
      type: object
      required: [sos_id, status, resolved_at]
      properties:
        sos_id:
          type: string
        status:
          type: string
          enum: [cancelled, safe]
        resolved_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Resource conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...